// src/app/models/attendancerequest.ts
export class AttendanceRequest {
    id: number;
    employeeId: number;
    loginTime: any;
    AloginTime: any;
    logoutTime: any;
    AlogoutTime: any;
    totalHours: any;
    totalBreak: number;
    comments: string;
    createdTime: string;
    modifieTime: string;
    tDate: any;
    date?: any;
    isActive: string;
    modifiedBy: number;
    createdBy: number;
}